Refactor ManagerMixin
